Moves dynakubeState into status package
