+++ src/controllers/dynakube/status/config.go
@@ -1,9 +1,9 @@
-package controllers
+package status
 
 import (
 	"github.com/Dynatrace/dynatrace-operator/src/logger"
 )
 
 var (
-	log = logger.NewDTLogger().WithName("controllers")
+	log = logger.NewDTLogger().WithName("dynakube-status")
 )
